Change item `max_uses` and `remaining_uses` from u8 to u16

(cherry picked from commit b01f2857d411a1685d466a3244b16cba3e54ba10)

+++ src/xrGame/script_game_object.h
@@ -1039,9 +1039,9 @@ public:
 	void SetArtefactAdditionalInventoryWeight(float value);
 
 	//Eatable items
-	void SetRemainingUses(u8 value);
-	u8 GetRemainingUses();
-	u8 GetMaxUses();
+	void SetRemainingUses(u16 value);
+	u16 GetRemainingUses();
+	u16 GetMaxUses();
 
 	//Phantom
 	void PhantomSetEnemy(CScriptGameObject*);
